🐛 fix(spinner): Fix test case for spinning functionality
The test case for spinning functionality (`test_spin`) in the `SpinnerState` module was not properly implemented. It was failing to correctly handle the synchronization of the spinning thread. To fix this, the test case has been modified to create a separate thread for the spinning logic using the `thread::spawn` function.

Changes made:
- Added `use std::thread;` to import the necessary module.
- Modified the `test_spin` test case to `test_spin_thread`.
- Created a new `spinner_thread` using `thread::spawn` to run the spinning logic in a separate thread.
- Modified the assertion to ensure that the spinning logic completes successfully.
- Added a line to stop the spinner by setting `running` to `false` after the spinning logic completes.
- Joined the `spinner_thread` to wait for it to finish execution.

These changes ensure proper synchronization and testing of the spinning functionality.

Signed-off-by: Sabry Awad <dr.sabry1997@gmail.com>