✨ feat(LICENSE): Add MIT license file
Created a new file `LICENSE` and added the standard MIT license text.

The license grants permission to any person to use, copy, modify, merge, publish, distribute, sublicense, and/or sell the software and associated documentation files free of charge, subject to certain conditions.

The conditions include including the copyright notice and permission notice in all copies or substantial portions of the software.

The software is provided "as is" without warranty of any kind.

Signed-off-by: Sabry Awad <dr.sabry1997@gmail.com>